move ANSI code to internal/ansi
