--- color_nonwindows.go
@@ -1,7 +1,0 @@
-//go:build !windows
-
-package readline
-
-func enableANSI() error {
-	return nil
-}
