+++ internal/ansi/ansi.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package ansi
+
+func EnableANSI() error {
+	return nil
+}
